this is great, lets customise the form for the Work order and then apply the same fields and designs to work order template.
1. header section
Business name, business address, business contact number and email.
in the right side of the header section option for user to add a business logo.

2. Work order details [table format]
Order number, order date, expected start date, expected end date

3. Client details [table format]
Client name, client phone, client email, work location, order received by

4. Work description - here user should be able to add multiple lines in the form
5. Terms of service - here user should be able to add multiple lines in the form

6. Work description table
It should be a table design with the following column names - Work description, Area in Sq. ft., Rate, Amount
The user should be able to add multiple lines in the form and same should be populated in the final design, amount will be in INR (₹) and final amount will be calculated below the table

7. Material description table
It should be a table design with the following column names - Material name, Unit, Price per Unit, Amount
The user should be able to add multiple lines in the form and same should be populated in the final design, amount will be in INR (₹) and final amount will be calculated below the table

8. Labour description table
It should be a table design with the following column names - Team name, number of person, amount
The user should be able to add multiple lines in the form and same should be populated in the final design, amount will be in INR (₹) and final amount will be calculated below the table

9. Final amount section
Here we will have some other cost fields such as - tax rate, any other costs and the final amount
The final total amount will be the sum of all the total  amounts of the table along with tax applied (if given by user)

10. Footer section
Approved by - name
Date of approval

Note: all the table should be enabled of disabled by the user in the form. If the table section is disabled we will not need to show or add those details in the final template.